editoast: refactor and move simulation hash logic
